Remove buggy CMD_RESOURCE_QUEUE handling
Currently the CMD_RESOURCE_QUEUE handler is blocking, which means it
waits for the decode thread indefinitely until the frame is decoded.
This will cause deadlock when the decode thread encouters some error or
there's simply not enough data to decode a frame.

The old implementation is removed and replaced by a skeleton of the new
implementation.

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>