Introduce the conception of VirtIOVideoWork
A work represents an on fly CMD_RESOURCE_QUEUE request. An input work
must first go to pending work queue, waiting for being submitted to
MediaSDK. After the request is accepted by MediaSDK, it is moved to
queued work queue, until the completion of decoding/encoding work.

A skeleton of CMD_RESOURCE_QUEUE is also provided.

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>