CMD_STREAM_DRAIN: handle init case more carefully
Doing STREAM_DRAIN while in STATE_INIT is tricky. It is good to support
this as a feature, but we should take care that if the stream cannot
complete the initialization phase and move on to the next phase, it is
meaningless to do a STREAM_DRAIN.

Previously we just stuck forever in this case, and now we gracefully
cancel the STREAM_DRAIN command.

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>