Fix pixel format handling
It turns out that ARGB8888 and BGRA8888 are packed formats, not planar
formats.

It seems that the channel order of MFX_FOURCC_RGB4 is B,G,R,A, at least
the sample code of MediaSDK use this order. But that sample code also
says that MFX_FOURCC_RGB4 is equivalent to VA_FOURCC_ARGB, whose
documentation says the channel order is A,R,G,B, which is very
confusing. Now we just use B,G,R,A order and see if it works.

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>